Progresso no método de exibir o resultado da requisição na tela de produtos

+++ App/app/src/main/java/com/teste/projeto_3/model/ResultadoRequisicaoEstoque.kt
@@ -25,7 +25,7 @@ class ResultadoRequisicaoEstoque(
     var descricaoProduto: String = "",
     var dataValidadeProduto: String = "",
     var quantidadeProduto: String = "",
-    var precoProduto: Float = 0f,
+    var precoProduto: Double = 0.0,
 ) {
     override fun toString(): String {
         return "ResultadoRequisicaoEstoque(nomeEstoque='$nomeEstoque', descricaoEstoque='$descricaoEstoque', textoResultado='$tituloResultado', iconeResultado=$iconeResultado, status='$status', descricaoStatus='$descricaoStatus', metodoDeEnvio='$metodoDeEnvio', imagemUrl='$imagemUrl', imagemBase64='$imagemBase64', idEstoque='$idEstoque', positionEstoqueLista='$positionEstoqueLista', idProduto='$idProduto', nomeProduto='$nomeProduto', descricaoProduto='$descricaoProduto', dataValidadeProduto='$dataValidadeProduto', quantidadeProduto='$quantidadeProduto', precoProduto='$precoProduto', positionProdutoLista='$positionProdutoLista')"
